Propuesta de actualización al código. Programación Object oriented.
